Split plotting into sub-module. Hookup Scatter with latest
