+++ geoapps/plotting/__init__.py
@@ -1,0 +1,8 @@
+from .plotting import (
+    normalize,
+    format_axis,
+    plot_plan_data_selection,
+    plot_profile_data_selection,
+)
+from .scatter import ScatterPlots
+from .slice import PlotSelection2D
